Fix breakage introduced by mitigation of CVE-2015-9284 (#41, #42)
90440b2bd60d broke saving of the user's swap preferences on login,
since an <a method="post" ...> element inside a <form> does not
include the other parameters within the <form> when submitting the
action via POST.  So replace it with a <button> element which does.
Also, since we want a single login modal which can work from any page
on the site, move the <form> into application.html.haml, give it an
id="login-form" attribute, and give the <select> elements a
form="login-form" attribute so that their values get submitted to it
on login.

However since we now have to POST directly to /auth/:provider in order
to mitigate CVE-2015-9284, we need some omniauth trickery to capture
the user's voting preferences before that POST, in lieu of the
existing require_login_and_save_user_params filter.  (The latter
previously worked because we used to initiate login via POST to our
own user_path.)

The new flow is:

- POST /auth/:provider, potentially including parameters
  detailing user's voting preferences

- omniauth before_request_phase handler saves those user preferences
  in the rack session (env["rack.session"])

- omniauth request phase calls the identity provider and
  (assuming success) gets identity information back

- omniauth callback phase calls session#create

- session#create gets or creates a User object, updates it with any
  user preferences retrieved from the session object, and redirects to
  whichever page the login request originated from, or user#show by
  default.

Fixes #41
Fixes #42
See also #36 although that is not fixed yet