Handle absence of user profile URL better in UI
Don't attempt to create links to the user's profile if we don't have
the URL.

There is no guarantee a user logging in via Facebook will allow us
permission to get their timeline (profile) link.  In fact if they've
previously authorised the app before we started asking for the
user_link permission in Omniauth, it might not even ask them whether
it's OK to share that link - Facebook seems flakey like that.

So make sure we can survive gracefully without the profile link.