Clarify the matching process
Ditch the oscillation of the fa-search magnifying glass, and replace
the rotating spinner with a static hourglass.

The animations made it look a bit like the views could automatically
update, leading people to thinking they just have to sit there waiting
for it.  As a result we needed a FAQ entry explaining it.  So ditch
the animation, as cute as it was :'(

Closes #71.