Clarify policy on use of email address
We already explain this in the Privacy Policy, but it could be
clearer.  In particular since we are now introducing email login which
has no social profile link, we need to make it explicit that in that
case we will share email address with the swap partner.

Also update the FAQ accordingly.

Addresses #102 at least partially.