Fix crashes when authentication fails
We've been seeing crashes reported in Airbrake with `Could not find a valid mapping for path "/auth/xxxxxx/callback"`. This seems to happen when authentication fails, e.g.:

* CSRF detection on the FB (Oauth2) logins because the `omniauth.state` in the session doesn't match the state passed back in the callback
* Failed login, e.g. login using Twitter but the user cancels the login
* Cancelled FB/Twitter app permission, so the user decides not to give our app permission to access their account

The problem seems to be that Devise can't figure out the route in this situation. To fix this we add a custom `on_failure` callback to ensure the right thing gets called.

See https://github.com/plataformatec/devise/issues/2004#issuecomment-7466322 for more information.