better name: Party#smv_code
We've been using this to uniquely identify parties for a long time, give
it a name that better explains. It's OUR way of identifying a party,
regardless of changes in database id or any external reference source.

There's a github issue for this I'm sure but I can't find it.