Improve sign-up flow
This removes the "choice" dialog from the flow as we only support email now. We get data processing consent on the email sign-up form now.

We add "required" stars on all the forms that have required fields.

We also disable the timeout on the captcha as it can cause problems when the user clicks submit multiple times.