Fix issue with missing emails
If the user doesn't have a verified mobile number and they try to swap then we send the validation code to them. Unfortunately if they don't have a number at all then we will crash as there's no mobile number to send the code to!

To fix this, if the user has no mobile number at all we redirect to the user edit page for them to provide it.

We also add a check that the email and name aren't the same, this fixes #953.

Finally we tidy up some more UI:

1. Adding required stars to things that are required
2. Centre buttons so we're consistent
3. Remove "!" from UI text

Closes #953
Closes #952